+++ src/state/persisted/index.ts
@@ -8,6 +8,7 @@ import {
   tryStringify,
 } from '#/state/persisted/schema'
 import {PersistedApi} from './types'
+import {normalizeData} from './util'
 
 export type {PersistedAccount, Schema} from '#/state/persisted/schema'
 export {defaults} from '#/state/persisted/schema'
@@ -19,7 +20,7 @@ let _state: Schema = defaults
 export async function init() {
   const stored = await readFromStorage()
   if (stored) {
-    _state = normalizeData(stored)
+    _state = stored
   }
 }
 init satisfies PersistedApi['init']
@@ -81,44 +82,9 @@ async function readFromStorage(): Promise<Schema | undefined> {
     })
   }
   if (rawData) {
-    return tryParse(rawData)
+    const parsed = tryParse(rawData)
+    if (parsed) {
+      return normalizeData(parsed)
+    }
   }
-}
-
-function normalizeData(data: Schema) {
-  /**
-   * Normalize language prefs to ensure that these values only contain 2-letter
-   * country codes without region.
-   */
-  try {
-    const next = {...data.languagePrefs}
-    next.primaryLanguage = next.primaryLanguage.split('-')[0]
-    next.contentLanguages = next.contentLanguages.map(lang =>
-      normalizeLocaleToTwoLetterCode(lang),
-    )
-    next.postLanguage = next.postLanguage
-      .split(',')
-      .map(lang => normalizeLocaleToTwoLetterCode(lang))
-      .filter(Boolean)
-      .join(',')
-    next.postLanguageHistory = next.postLanguageHistory.map(postLanguage => {
-      return postLanguage
-        .split(',')
-        .map(lang => normalizeLocaleToTwoLetterCode(lang))
-        .filter(Boolean)
-        .join(',')
-    })
-    // mutate last in case anything above fails
-    data.languagePrefs = next
-  } catch (e: any) {
-    logger.error(`persisted state: failed to normalize language prefs`, {
-      safeMessage: e.message,
-    })
-  }
-
-  return data
-}
-
-function normalizeLocaleToTwoLetterCode(lang: string) {
-  return lang.split('-')[0]
 }
